+++ routes/press.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
+use Orchid\Press\Http\Controllers\MenuController;
 use Orchid\Press\Http\Screens\EntityEditScreen;
 use Orchid\Press\Http\Screens\EntityListScreen;
-use Orchid\Press\Http\Controllers\MenuController;
 
 /*
 |--------------------------------------------------------------------------
@@ -22,13 +22,13 @@ $this->router->screen('entities/{type}', EntityListScreen::class)->name('entitie
 
 $this->router->resource('menu', MenuController::class, [
     'only'  => [
-        'index', 'show', 'update', 'create', 'destroy',
+        'index', 'show', 'update', 'store', 'destroy',
     ],
     'names' => [
         'index'   => 'systems.menu.index',
         'show'    => 'systems.menu.show',
         'update'  => 'systems.menu.update',
-        'create'  => 'systems.menu.create',
+        'store'   => 'systems.menu.store',
         'destroy' => 'systems.menu.destroy',
     ],
 ]);